join button callback function

+++ src/main/resources/db/v_1.0/01_create_users_table.sql
@@ -2,8 +2,8 @@ USE jointbuy ;
 
 CREATE TABLE Members (
   ID INT NOT NULL AUTO_INCREMENT,
-  telegramUserId INT,
-  telegramChatId INT,
+  telegramUserId DECIMAL(15),
+  telegramChatId DECIMAL(15),
   firstName VARCHAR(255) NOT NULL,
   lastName VARCHAR(255),
   alias VARCHAR(255),
